+++ NEWS.md
@@ -44,6 +44,7 @@
 
 ## Other new features
 
+* Updated vendored Catch framework from v1.9.6 (2017) to Catch2 v2.13.10 (2022). This fixes compilation issues with the C++23 standard. @pachadotdev
 * New `extract_test()` function to extract a reprex from a failing expectation. tests run in `R CMD check` will use this to automatically create a reprex in the `_problems/` directory for each failing expectation. You can turn this behaviour off by setting `TESTTHAT_PROBLEMS=false` (#2263).
 * New `local_mocked_s3_method()`, `local_mocked_s4_method()`, and `local_mocked_r6_class()` allow you to mock S3 and S4 methods and R6 classes (#1892, #1916)
 * New `local_on_cran(TRUE)` allows you to simulate how your tests will run on CRAN (#2112).